+++ src/app/modules/feedback/__tests__/feedback.service.spec.ts
@@ -8,7 +8,7 @@ import getFormFeedbackModel from 'src/app/models/form_feedback.server.model'
 
 import dbHandler from 'tests/unit/backend/helpers/jest-db'
 
-import { GetFormFeedbackDto } from '../../../../types/api/form-feedback'
+import { GetFormFeedbackDto } from '../../../../types/api/form_feedback'
 import { DatabaseError } from '../../core/core.errors'
 import * as FeedbackService from '../feedback.service'
