+++ src/app/modules/feedback/feedback.service.ts
@@ -4,7 +4,7 @@ import mongoose from 'mongoose'
 import { ResultAsync } from 'neverthrow'
 
 import { IFormFeedbackSchema } from '../../../types'
-import { GetFormFeedbackDto } from '../../../types/api/form-feedback'
+import { GetFormFeedbackDto } from '../../../types/api/form_feedback'
 import { ProcessedFeedback } from '../../../types/form_feedback'
 import { createLoggerWithLabel } from '../../config/logger'
 import getFormFeedbackModel from '../../models/form_feedback.server.model'
