+++ src/public/modules/forms/helpers/FeedbackCsvGenerator.ts
@@ -1,6 +1,6 @@
 import moment from 'moment-timezone'
 
-import { FormFeedbackResponseDto } from '../../../../types/api/form-feedback'
+import { FormFeedbackResponseDto } from '../../../../types/api/form_feedback'
 
 import { CsvGenerator } from './CsvGenerator'
