+++ src/types/api/index.ts
@@ -3,6 +3,6 @@ export * from './auth'
 export * from './form'
 export * from './billing'
 export * from './examples'
-export * from './encrypt-submission'
+export * from './encrypt_submission'
 export * from './email-submission'
 export * from './submission'
